pyproject.toml
[project]
name = "phivolcs-etl-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.1",
    "lxml>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.42",
    "urllib3>=2.5.0",
]
